+++ packages/core/contracts/test/Storage.sol
@@ -313,6 +313,20 @@ contract StorageUpgrade_ObviousMismatch_V2_Bad {
     mapping (uint => uint) a1;
 }
 
+contract StorageUpgrade_FunctionType_Visibility_V1 {
+    struct S {
+        function () internal m;
+    }
+    S s1;
+}
+
+contract StorageUpgrade_FunctionType_Visibility_V2_Bad {
+    struct S {
+        function () external m;
+    }
+    S s1;
+}
+
 contract StorageUpgrade_Contract_V1 {
     StorageUpgrade_Contract_V1 data;
 }