+++ packages/core/src/storage/compare.ts
@@ -28,6 +28,7 @@ export type TypeChange = (
         | 'obvious mismatch'
         | 'unknown'
         | 'array grow'
+        | 'visibility change'
         | 'array shrink'
         | 'array dynamic'
         | 'enum resize'
@@ -78,6 +79,16 @@ export class StorageLayoutComparator {
     }
   }
 
+  getVisibilityChange(original: ParsedTypeDetailed, updated: ParsedTypeDetailed): TypeChange | undefined {
+    const re = /^t_function_(internal|external)/;
+    const originalVisibility = original.head.match(re);
+    const updatedVisibility = updated.head.match(re);
+    assert(originalVisibility && updatedVisibility);
+    if (originalVisibility[0] !== updatedVisibility[0]) {
+      return { kind: 'visibility change', original, updated };
+    }
+  }
+
   getFieldChange<F extends StorageField>(original: F, updated: F): StorageFieldChange<F> | undefined {
     const nameChange = !this.unsafeAllowRenames && original.label !== updated.label;
     const typeChange = this.getTypeChange(original.type, updated.type, { allowAppend: false });
@@ -121,6 +132,10 @@ export class StorageLayoutComparator {
     updated: ParsedTypeDetailed,
     { allowAppend }: { allowAppend: boolean },
   ): TypeChange | undefined {
+    if (updated.head.startsWith('t_function')) {
+      return this.getVisibilityChange(original, updated);
+    }
+
     if (original.head !== updated.head) {
       return { kind: 'obvious mismatch', original, updated };
     }